requirements.txt
Python==3.12
LangChain>=0.3.25
langchain-deepseek>=0.1.3
langchain-community>=0.3.0
python-dotenv
openai
langchain-deepseek
gradio
# pip install playwright lxml langchain_community beautifulsoup4 reportlab